backend/src/app.py
"""
Main Flask application
"""
import os
from flask import Flask, jsonify
from flask_cors import CORS
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Create Flask app
app = Flask(__name__)

# Configure CORS
CORS(app, origins=os.getenv("ALLOWED_ORIGINS", "*").split(","))

# Configuration
app.config["DEBUG"] = os.getenv("FLASK_DEBUG", "False").lower() == "true"
app.config["SECRET_KEY"] = os.getenv("SECRET_KEY", "dev-secret-key")


@app.route("/")
def index():
    """Root endpoint"""
    return jsonify({
        "message": "Flask API is running",
        "version": "0.1.0",
        "environment": os.getenv("ENVIRONMENT", "development")
    })


@app.route("/api/health")
def health():
    """Health check endpoint"""
    return jsonify({
        "status": "healthy",
        "service": "backend-api"
    })


if __name__ == "__main__":
    port = int(os.getenv("PORT", 5000))
    app.run(host="0.0.0.0", port=port, debug=app.config["DEBUG"])